backend/requirements.txt
Flask
flask-cors
google-api-python-client